Add client prefix registration with server
This change adds the first steps towards autoconfiguration in networks with communication restrictions. When the client app is run, it sends the prefixes it serves to the server. For now, we get the prefixes by asking the user but a TODO item is to have the client code get prefixes in a better way.

The server accepts this message and handles the registration request by saving the list of prefixes and the IP address. It saves these in the same format (ndn::Block, which each encode one TLV) that they were sent in because we have to send this mapping back to clients using the same format.

Client.cpp
For now, just initializes registration

RegistrationRequest_client.cpp
Asks client for prefixes and each time they enter one, it saves the prefix as a TLV (type=prefix) and adds it to the list of prefixes TLV (type = PrefixListToRegister). It then encodes the list of prefixes and sends to RV

Server.cpp
Waits for messages from clients. When it receives one, it parses the wire to retrieve an ndn::Block (see ndn-cxx/encoding for more info). For now, it ignores all message types except a registration request, for which it hands off to handleRegistrationRequest

RegistrationRequest_server.cpp
Stores the list of prefixes and the IP address in a mapping and adds the mapping to its static list (also a TLV). For now, this does not check whether the IP address already exists. A TODO item is to decide whether/how we want updates to work