Add RV response to prefix registration
When RV receives a list of prefixes to register, it sends back all the IP address – prefix mappings it knows about so that nodes can unicast going forward (we are still assuming multicast is not allowed and unicast is allowed). When RV received the list of prefixes, it saved them in this format (all are TLVs for ease of sending back to client). In this example, RV knows about two nodes, each of which store two prefixes.
IPPrefixMappingList
	IPPrefixMapping
		IPAddress
		PrefixListToRegister
			Prefix
			Prefix
	IPPrefixMapping
		IPAddress
		PrefixListToRegister
			Prefix
			Prefix

RegistrationRequest_client.cpp
Separated out the method to get prefixes. Added logic to receive response from server after sending. If the type is the expected IPPrefixMappingList, send it to a helper method, parseMappingListBlock, which for now just prints the contents to verify.

RegistrationRequest_server.cpp
Separate the method to add the prefix list to the mapping. Add logic to encode and send the mapping list upon receiving a registration request.

Note: with this logic, the mapping list sent back to client includes the prefixes of the client we just received from. We could find a way to get rid of that in the future if necessary, but we don’t want to risk not storing those prefixes in the case where something goes wrong in the response