Store IP-prefix mappings from server in local NFD
This change adds logic to take the IP address - prefix list mapping received by RV and turn it into fib entries using control commands. RV sends back a list of nodes it knows about and the prefixes they each publish under. With this change,for each node (IP address) in the mapping, the client creates a face, and for each prefix associated with it, the client creates a fib entry, adding the the UDP-IP tunnel as the next hop for each prefix

Note: this change assumes that each client registers with RV exactly once (that is, there is currently no checking for whether information about the node already exists in the local NFD)