Add support for IPv6
We were using all the C++ support for IPv4 alone (the methods and structs designed to hold and convert IPv4 addresses). C++ has support for IPv6, which for the most part means using sockaddr_storage instead of sockaddr_in and using getaddrinfo and getnameinfo wherever possible. An added bonus of getaddrinfo is that when we have the name of the RV instead of just the hard-coded IP address, we can stick that in and it will automatically do DNS lookup.