Enable clients to send their own IP addresses to RV
Before, each client would send the prefixes it publishes under to RV and RV would take the sender's IP address and associate it with the client. If the client is behind a NAT though, the address that RV sees is meaningless for communicating within the network, which is what we want clients to be able to do. Instead, let's have each client send their own local IP along with the prefixes. That also makes it easier for each client to exclude itself when adding entries to the fib.

RegistrationRequest_client.cpp: added a method to retrieve local IP (needs editing once I figure out how to do this in a cross-platform way) and add the local IP to the prefix list before sending. When storing a new entry in the fib, first check that it's not the local IP

RegistrationRequest_server.cpp: removed all the logic for retrieving and storing the IP of the client because now that's part of the data.